// ui/src/api/generated/model/browserPushSubscription.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Whirlpool PDI Tool API
 * Whirlpool PDI Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.7.1
 */
import type { PushKeys } from './pushKeys';

export interface BrowserPushSubscription {
  /**
     * @minLength 1
     * @maxLength 2048
     */
  endpoint: string;
  expirationTime?: number | null;
  keys: PushKeys;
}