// ui/src/api/generated/model/pushSubscriptionCreate.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Whirlpool PDI Tool API
 * Whirlpool PDI Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.6.0
 */
import type { BrowserPushSubscription } from './browserPushSubscription';

export interface PushSubscriptionCreate {
  subscription: BrowserPushSubscription;
  device_uuid?: string | null;
  device_fingerprint?: string | null;
  /** @maxLength 2000 */
  user_agent?: string;
  /** @maxLength 64 */
  timezone?: string;
}