// ui/src/api/generated/model/pushUserSendRequest.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Whirlpool PDI Tool API
 * Whirlpool PDI Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.5.1
 */
import type { PushSendPayload } from './pushSendPayload';

export interface PushUserSendRequest {
  user_uuid: string;
  notification: PushSendPayload;
}